A great many improvements
Added GAME_OVER state to clients, not working yet due to the state being instantly changed to queueing by server.

Changed the game started text to show who is vs who in the server console.

When a game finishes, queued players are placed into a new game, if there are enough.

Various fixes and improvements.